+++ flint/validation.py
@@ -18,6 +18,7 @@ from matplotlib.axes import Axes
 from matplotlib.figure import Figure
 from scipy import stats
 
+from flint.catalogue import Catalogue, get_reference_catalogue
 from flint.logging import logger
 from flint.naming import processed_ms_format
 from flint.summary import BeamSummary, FieldSummary
@@ -29,37 +30,9 @@ F_LARGE = 12
 F_HUGE = 20
 
 
-class Catalogue(NamedTuple):
-    """A basic structure used to describe a known catalogue."""
-
-    survey: str
-    """Shorthand name of the sourcey catalogue"""
-    file_name: str
-    """The file name of the known catalogue"""
-    freq: float  # Hertz
-    """Reference frequency of the catalogue, in Hertz"""
-    ra_col: str
-    """Column name containing the right-ascension"""
-    dec_col: str
-    """Column name containing the declination"""
-    name_col: str
-    """Column name containing the source/component name"""
-    flux_col: str
-    """Column name containing the flux density"""
-    maj_col: str
-    """Column name containing the major-axis of the source gaussian component"""
-    min_col: str
-    """Column name containing the min-axis of the source gaussian component"""
-    pa_col: str
-    """Column name containing the pa of the source gaussian component"""
-    alpha_col: Optional[str] = None  # Used to scale the SED
-    """Column name containing the spectral index, used to calculate the source SED. If None a default is used. """
-    q_col: Optional[str] = None  # Used to scale the SED
-    """Column name containing the curvature of the spectral index, used to calculate the source SED. If None a default is used. """
-
-
-class Catalogues(NamedTuple):
-    """Container for all the catalogues that are loaded in"""
+class ValidationCatalogues(NamedTuple):
+    """Container for all the catalogues that are loaded in and
+    used throughout validation processing"""
 
     nvss: Catalogue
     """NVSS catalogue"""
@@ -312,17 +285,9 @@ def load_known_catalogue(
     Returns:
         Tuple[Table,Catalogue]: The loaded table and Catalogue structure describing the columns
     """
-    catalogue = get_known_catalogue_info(name=name)
-    catalogue_path = reference_catalogue_directory / catalogue.file_name
-    table = Table.read(catalogue_path)
-
-    if name == "SUMSS":
-        table[catalogue.flux_col] = table[catalogue.flux_col] * u.mJy
-    if name == "ICRF":
-        return table, catalogue
-
-    table[catalogue.flux_col] = table[catalogue.flux_col].to(u.Jy).value
-
+    table, catalogue = get_reference_catalogue(
+        reference_directory=reference_catalogue_directory, survey=name
+    )
     return table, catalogue
 
 
@@ -1101,7 +1066,7 @@ def load_catalogues(
     reference_catalogue_directory: Path,
     askap_survey_name: str,
     rms_info: RMSImageInfo,
-) -> Tuple[Catalogues, Tables]:
+) -> Tuple[ValidationCatalogues, Tables]:
     """Load in all the catalogues that are required for the validation.
 
     Args:
@@ -1111,7 +1076,7 @@ def load_catalogues(
         rms_info (RMSImageInfo): The extracted information from the RMS image
 
     Returns:
-        Tuple[Catalogues, Tables]: The loaded catalogues and tables
+        Tuple[ValidationCatalogues, Tables]: The loaded catalogues and tables
     """
     logger.info(f"Loading {source_catalogue_path=}")
     askap_table = Table.read(source_catalogue_path)
@@ -1138,7 +1103,7 @@ def load_catalogues(
     )
 
     return (
-        Catalogues(
+        ValidationCatalogues(
             askap=askap_cata,
             icrf=icrf_catalogue,
             sumss=sumss_catalogue,